move eth rpc logic to separate module from controller
