+++ apps/block_scout_web/lib/block_scout_web/controllers/api_docs_controller.ex
@@ -1,8 +1,8 @@
 defmodule BlockScoutWeb.APIDocsController do
   use BlockScoutWeb, :controller
 
-  alias BlockScoutWeb.API.RPC.EthController
   alias BlockScoutWeb.Etherscan
+  alias Explorer.EthRPC
 
   def index(conn, _params) do
     conn
@@ -12,7 +12,7 @@ defmodule BlockScoutWeb.APIDocsController do
 
   def eth_rpc(conn, _params) do
     conn
-    |> assign(:documentation, EthController.methods())
+    |> assign(:documentation, EthRPC.methods())
     |> render("eth_rpc.html")
   end
 end
